jujutracer documentation

Documentation for the package jujutracer.

Basic usage

Convert a PFM file into a LDR format such as .png or .jpg. Need $a$ and $\gamma$ values specified.

julia main.jl <pfm_file> <a> <gamma> <output_file>

Module Index

jujutracer.AbstractTransformation

jujutracer.AbstractTransformation

jujutracer.AbstractTransformation

jujutracer.AbstractTransformation

jujutracer.InvalidPfmFileFormat

jujutracer.Normal

jujutracer.Point

jujutracer.Rx

jujutracer.Ry

jujutracer.Rz

jujutracer.Scaling

jujutracer.Transformation

jujutracer.Translation

jujutracer.Vec

jujutracer.hdrimg

jujutracer.:×

jujutracer.:⊙

jujutracer.:⋅

jujutracer._RGBluminosity

jujutracer._average_luminosity

jujutracer._clamp_img!

jujutracer._lumi_D

jujutracer._lumi_Func

jujutracer._lumi_mean

jujutracer._lumi_weighted

jujutracer._normalize_img!

jujutracer._parse_endianness

jujutracer._parse_image_size

jujutracer._read_float

jujutracer._read_line

jujutracer._write_float!

jujutracer._γ_correction!

jujutracer.get_matrix

jujutracer.inverse

jujutracer.norm

jujutracer.normalize

jujutracer.read_pfm_image

jujutracer.save_ldrimage

jujutracer.squared_norm

jujutracer.to_string

jujutracer.tone_mapping

jujutracer.write_pfm_image

Detailed API

jujutracer.AbstractTransformation — Type

AbstractTransformation

An abstract type that serves as a base for defining various geometric transformations. Made concrete by Transformation, Translation, Scaling, Rx, Ry, and Rz.

jujutracer.AbstractTransformation — Method

(t::AbstractTransformation)(n::Normal)

Applies the transformation to a Normal.

jujutracer.AbstractTransformation — Method

(t::AbstractTransformation)(p::Point)

Applies the transformation to a Point.

jujutracer.AbstractTransformation — Method

(t::AbstractTransformation)(v::Vec)

Applies the transformation to a Vec.

jujutracer.InvalidPfmFileFormat — Type

InvalidPfmFileFormat <: Exception

Exception raised when the PFM file format is invalid or cannot be read.

Fields

error_message::String: The error message describing the issue.

jujutracer.Normal — Type

Normal(x::Float64, y::Float64, z::Float64)

Struct representing a unit vector (normal) in 3D space.

Fields

x::Float64,y::Float64,z::Float64: Coordinates.

Methods

Normal(v::Vec): Create a Normal from a Vec.

Normal(x::Float64, y::Float64, z::Float64): Create a Normal from x, y, z coordinates.

Throws

ArgumentError: If the vector is zero.

jujutracer.Point — Type

Point(x::Float64, y::Float64, z::Float64)

Struct representing a point in 3D space.

Fields

x::Float64,y::Float64,z::Float64: Coordinates.

jujutracer.Rx — Type

struct Rx <: AbstractTransformation

Represents a rotation transformation around the x-axis in 3D space. This structure is a subtype of AbstractTransformation.

Fields

M::Matrix{Float64}: The 4x4 transformation matrix.

inv::Matrix{Float64}: The 4x4 inverse of the transformation matrix.

Constructor

Rx(θ): Creates an Rx instance for a given rotation θ (in radians).

See also

Ry: For rotation around the y-axis.

Rz: For rotation around the z-axis.

jujutracer.Ry — Type

struct Ry <: AbstractTransformation

Represents a rotation transformation around the y-axis in 3D space. This structure is a subtype of AbstractTransformation.

Fields

M::Matrix{Float64}: The 4x4 transformation matrix.

inv::Matrix{Float64}: The 4x4 inverse of the transformation matrix.

Constructor

Ry(θ): Creates an Ry instance for a given rotation θ (in radians).

See also

Rx: For rotation around the x-axis.

Rz: For rotation around the z-axis.

jujutracer.Rz — Type

struct Rz <: AbstractTransformation

Represents a rotation transformation around the z-axis in 3D space. This structure is a subtype of AbstractTransformation.

Fields

M::Matrix{Float64}: The 4x4 transformation matrix.

inv::Matrix{Float64}: The 4x4 inverse of the transformation matrix.

Constructor

Rz(θ): Creates an Rz instance for a given rotation θ (in radians).

See also

Rx: For rotation around the x-axis.

Ry: For rotation around the y-axis.

jujutracer.Scaling — Type

struct Scaling <: AbstractTransformation

Represents a scaling transformation in 3D space. This structure is a subtype of AbstractTransformation.

Fields

M::Matrix{Float64}: The 4x4 transformation matrix.

inv::Matrix{Float64}: The 4x4 inverse of the transformation matrix.

Constructor

Scaling(x::Float64, y::Float64, z::Float64): Creates aScalinginstance with scaling factorsx,y, andzalong the x, y, and z axes, respectively. Throws anArgumentError` if any of the scaling factors are zero.

jujutracer.Transformation — Type

struct Transformation <: AbstractTransformation

Represents a transformation in 3D space with homogeneous coordinates. This structure is a subtype of AbstractTransformation.

Fields

M::Matrix{Float64}: The 4x4 transformation matrix.

inv::Matrix{Float64}: The 4x4 inverse of the transformation matrix.

Constructors

Transformation(): Creates an identity transformation where M and inv are both 4x4 identity matrices.

Transformation(M::Matrix{Float64}, inv::Matrix{Float64}): Creates a transformation with the given M and inv matrices.

Throws an ArgumentError if:

M or inv are not 4x4 matrices.

The last element of M or inv is not 1.0.

M and inv are not inverses of each other.

Notes

The M and inv matrices must satisfy the following conditions:

Both must be 4x4 matrices.

The last element of both matrices must be 1.0.

The product of M and inv must be approximately equal to the 4x4 identity matrix.

jujutracer.Translation — Type

struct Translation <: AbstractTransformation

Represents a translation in 3D space with homogeneous coordinates. This structure is a subtype of AbstractTransformation.

Fields

M::Matrix{Float64}: The 4x4 transformation matrix.

inv::Matrix{Float64}: The 4x4 inverse transformation matrix.

Constructors

Translation(dx::Float64, dy::Float64, dz::Float64): Creates a Translation object with translation offsets dx, dy, and dz along the x, y, and z axes, respectively.

Translation(v::Vec): Creates a Translation object using a Vec object.

jujutracer.Vec — Type

Vec(x::Float64, y::Float64, z::Float64)

A struct representing a vector in 3D space.

Fields

x::Float64,y::Float64,z::Float64: Coordinates.

Methods

Vec(n::Normal): Create a Vec from a Normal.

Vec(p::Point): Create a Vec from a Point.

jujutracer.hdrimg — Type

hdrimg(w::Int, h::Int)

A struct representing a high dynamic range image (HDR image).

Fields

img::Matrix{RGB}: A matrix of RGB values representing the HDR image.

w::Int64: The width of the image in pixels.

h::Int64: The height of the image in pixels.

Base.:≈ — Method

≈(a::Union{Transformation, Translation, Scaling, Rx, Ry, Rz}, b::Union{Transformation, Translation, Scaling, Rx, Ry, Rz})

Defines an approximate equality operator ≈ for geometric transformations. Two transformations a and b are considered approximately equal if both their transformation matrices (M) and their inverses (inv) are approximately equal.

jujutracer.:× — Method

a × b

Return wedge product ($\times$) between Vec or Normal as a Vec.

jujutracer.:⊙ — Method

⊙(a, b)

Composition of two transformations, where b is the first acting on the object and a the second.

jujutracer.:⋅ — Method

a ⋅ b

Return scalar product ($\cdot$) between Vec or Normal.

jujutracer._RGBluminosity — Function

_RGBluminosity(color, type = "LF")

Calculate the luminosity of a color using different methods.

Arguments

color: The RGB color to be processed. Expected to be a struct of 3 values (r, g, b).

type: The type of luminosity calculation to perform. Options are:

'LF' : Luminosity function (default),

'M' : Mean luminosity,

'W' : Weighted luminosity,

'D' : Euclidean distance luminosity.

Returns

The luminosity value of the pixel.

jujutracer._average_luminosity — Method

_average_luminosity(img::hdrimg; type = "LF", delta = 0.0001)

Return the average luminosity of an HDR image, given the type of luminosity calculation to be used and delta to avoid.

Arguments

img::hdrimg: The HDR image for which to calculate the average luminosity.

type::String: The type of luminosity calculation to be used. Options are:

LF: Luminosity function (default)

M: Mean luminosity

W: Weighted luminosity

D: Euclidean distance luminosity

delta::Float64: A small value to avoid log(0) for black pixels (default is 0.0001).

Returns

Float64: The average luminosity of the HDR image.

jujutracer._clamp_img! — Method

_clamp_img!(hdr::hdrimg)

Clamp the HDR image values to the range [0, 1] using the formula:

\[R_i → \frac{R_i}{1+R_i}\]

Arguments

hdr::hdrimg: The HDR image to be clamped.

jujutracer._lumi_D — Method

_lumi_D(color)

Euclidean distance luminosity function.

Arguments

The RGB color to be processed. Expected to be a struct of 3 values (r, g, b).

Returns

The luminosity value of the pixel.

jujutracer._lumi_Func — Method

_lumi_Func(color)

Luminosity function according to the formula: $\frac{max(r,g,b) + min(r,g,b)}{2}$.

Arguments

The RGB color to be processed. Expected to be a struct of 3 values (r, g, b).

Returns

The luminosity value of the pixel.

jujutracer._lumi_mean — Method

_lumi_mean(color)

Basic mean luminosity function.

Arguments

The RGB color to be processed. Expected to be a struct of 3 values (r, g, b).

Returns

The mean luminosity value of the pixel.

jujutracer._lumi_weighted — Method

_lumi_weighted(color)

Weighted mean luminosity function according to CCIR 601.

Arguments

The RGB color to be processed. Expected to be a struct of 3 values (r, g, b).

Returns

The luminosity value of the pixel.

jujutracer._normalize_img! — Method

_normalize_img!(img::hdrimg: a::T, std::T) where {T<:Real, N}

Normalize an image by using

\[R_i → R_i × \frac{R_i}{⟨l⟩}\]

Arguments

img::hdrimg: The HDR image to be normalized.

a::T: A positive value to be used in the normalization formula.

lum::T: The average luminosity of the image. If not provided, it will be calculated using _average_luminosity.

Raises

ArgumentError: If a is not a positive number or if lum is not a number.

jujutracer._parse_endianness — Method

_parse_endianness(endian::String)

Parse endianness of the PFM file from a string.

Arguments

endian::String: The string representation of the endianness, expected to be "±1.0".

Returns

Bool: Returns true if the endianness is little-endian, false if big-endian.

Raises

InvalidPfmFileFormat: If the endianness value is invalid or missing.

jujutracer._parse_image_size — Method

_parse_image_size(line::String)

Read image size from a string.

Arguments

line::String: The string representation of the image size, expected to be a couple of ints "width height".

Returns

Tuple{Int, Int}: A tuple containing the width and height of the image.

Raises

InvalidPfmFileFormat: If the image size specification is invalid or if the width/height values are not integers.

jujutracer._read_float — Method

_read_float(io::IO, is_little_endian)

Read a Float32 value by interpreting 4 bytes with the correct endianness from a buffer.

Arguments

io::IO: The input stream from which to read the float.

is_little_endian::Bool: A boolean indicating whether the float is in little-endian format.

Returns

Float32: The read float value.

Raises

InvalidPfmFileFormat: If there is an error reading the bytes from the file.

jujutracer._read_line — Method

_read_line(io::IO)

Read a line from the input buffer.

Arguments

io::IO: The input stream from which to read the line.

Returns

String: The read line.

jujutracer._write_float! — Function

_write_float!(f, io, endianness::Bool=true)

Write a Float32 value to the output stream with the correct endianness.

Arguments

f: The float value to be written.

io: The output stream to which the float will be written.

endianness::Bool: A boolean indicating whether to write the float in little-endian format (default is true).

jujutracer._γ_correction! — Method

_γ_correction!(hdr::hdrimg; γ = 1.0)

Apply gamma correction to the HDR image using the formula:

\[R_i → {R_i}^{1/γ}\]

Arguments

hdr::hdrimg: The HDR image to be gamma corrected.

γ::Float64: The gamma value to be used for correction. Must be a positive number (default is 1.0).

Raises

ArgumentError: If γ is not a positive number.

jujutracer.get_matrix — Method

get_matrix(img::hdrimg)

Extract the matrix from an HDR image.

Arguments

img::hdrimg: The HDR image from which to extract the matrix.

Returns

Matrix: The matrix representation of the HDR image.

jujutracer.inverse — Method

inverse(a::AbstractTransformation)

Return the inverse transformation

jujutracer.norm — Method

norm(v::Union{Vec, Normal})

Calculates the norm of a vector.

Arguments

v::Union{Vec, Normal}: The vector to be processed. Expected to be a Vec or Normal of 3 values (x, y, z).

Returns

The norm of the vector.

jujutracer.normalize — Method

normalize(v::Union{Vec, Normal})

Return a normalized vector.

Arguments

v::Vec: The vector to be normalized.

Returns

The normalized vector.

Throws

ArgumentError: If the vector is zero.

jujutracer.read_pfm_image — Method

read_pfm_image(io::IO)

Read a PFM image from an input stream.

Arguments

io::IO: The input stream from which to read the PFM image.

Returns

hdrimg: The HDR image read from the PFM file.

Raises

InvalidPfmFileFormat: If the PFM file format is invalid or if there are issues reading the file.

jujutracer.save_ldrimage — Method

save_ldrimage(img_matrix::Matrix, filename::String)

Save an LDR image to a file.

Arguments

img_matrix::Matrix: The matrix representation of the image to be saved.

filename::String: The name of the file to save the image to, including the file extension (".png" or ".jpg").

Returns

String: The path to the saved file.

Raises

ArgumentError: If the file extension is not valid or if the image values are not clamped.

jujutracer.squared_norm — Method

squared_norm(v::Union{Vec, Normal})

Calculates the squared norm of a vector.

Arguments

v::Union{Vec, Normal}: The vector to be processed. Expected to be a struct of 3 values (x, y, z).

Returns

The squared norm of the vector.

jujutracer.to_string — Method

to_string(v::T) where {T<:Union{Point, Vec, Normal}}

Converts a Point, Vec, or Normal object to a string representation.

Arguments

v::T: The object to be converted. Expected to be a struct of 3 values (x, y, z).

Returns

A string representation of the object.

jujutracer.tone_mapping — Method

tone_mapping(img::hdrimg; a=0.18, lum = nothing, γ = 1.0)

Apply tone mapping to the HDR image.

Arguments

img::hdrimg: The HDR image to be tone-mapped.

a::Float64: A positive value to be used in the normalization formula (default is 0.18).

lum::Float64: The average luminosity of the image. If not provided, it will be calculated using _average_luminosity.

γ::Float64: The gamma value to be used for correction. Must be a positive number (default is 1.0).

Returns

hdrimg: The tone-mapped HDR image.

jujutracer.write_pfm_image — Function

write_pfm_image(img::hdrimg, io, endianness::Bool=true)

Write a PFM file encodiding the content of an hdrimg

Arguments

img::hdrimg: The HDR image to be written to the PFM file.

io::IO: The output stream to which the PFM image will be written.

endianness::Bool: A boolean indicating whether to write the float values in little-endian format (default is true).